+++ pytest_flask/fixtures.py
@@ -69,7 +69,7 @@ def live_server(request, app, pytestconfig):
     host = pytestconfig.getvalue("live_server_host")
 
     # Explicitly set application ``SERVER_NAME`` for test suite
-    original_server_name = app.config["SERVER_NAME"] or "localhost"
+    original_server_name = app.config["SERVER_NAME"] or "localhost.localdomain"
     final_server_name = _rewrite_server_name(original_server_name, str(port))
     app.config["SERVER_NAME"] = final_server_name